Removed i32 from terminal api, swapped y axis, updated readme - Functions that took i32 for coordinates now take usize to avoid having   to convert everywhere when indexing - Y=0 now refers to the top of the terminal. Makes more sense for how   text is written to  the console

+++ examples/noise.rs
@@ -116,7 +116,7 @@ fn noise(time: Res<Time>, mut noise: ResMut<Noise>, mut q: Query<(&mut Terminal,
 
         let t = noise.noise.get_noise_type();
         let string = to_string(t);
-        let h = term.height() as i32;
+        let h = term.height();
         term.clear_box(0, h - 1, string.len(), 1);
         term.put_string(0, h - 1, &string);
     }
